src/main/java/lv/javaguru/java1/student_artjoms_potasevskis/lesson_12/AppleWarehouse.java
package lv.javaguru.java1.student_artjoms_potasevskis.lesson_12;


import java.util.ArrayList;
import java.util.List;

class AppleWarehouse {

    private List<Apple> apples = getAllApples();

    public List<Apple> findGreenApples() {

        List<Apple> greenApples = new ArrayList<>();
        for (Apple apple : apples) {
            if (apple.getColor().equals("green")) {
                greenApples.add(apple);
            }
        }
        return greenApples;
    }
    public List<Apple> findRedApples() {

        List<Apple> redApples = new ArrayList<>();
        for (Apple apple : apples) {
            if (apple.getColor().equals("red")) {
                redApples.add(apple);
            }
        }
        return redApples;
    }
    public List<Apple> findYellowApples() {

        List<Apple> yellowApples = new ArrayList<>();
        for (Apple apple : apples) {
            if (apple.getColor().equals("yellow")) {
                yellowApples.add(apple);
            }
        }
        return yellowApples;
    }

    private List<Apple> getAllApples() {
        List<Apple> apples = new ArrayList<>();
        apples.add(new Apple("red", 100));
        apples.add(new Apple("red", 160));
        apples.add(new Apple("red", 200));
        apples.add(new Apple("green", 200));
        apples.add(new Apple("green", 90));
        apples.add(new Apple("green", 50));
        apples.add(new Apple("yellow", 50));
        apples.add(new Apple("yellow", 170));
        return apples;
    }

}
